pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeep"
version = "0.1.2"
description = "A Python package for codeep upgrade optimization"
readme = "README.md"
authors = [{name = "Codeep Team", email = "andyxiaolei@gmail.com"}]
license = "MIT"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "requests>=2.20.0",
    "langchain>=0.1.0",
    "pydantic>=1.0.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
    "mypy>=0.800",
]

[project.urls]
"Homepage" = "https://github.com/codeepcc/codeep"
"Repository" = "https://github.com/codeepcc/codeep"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # Exclude the docs directory
  docs
| src/codeep_upgrade/_version.py
)/
'''